move autotun patch to xtuner.v1 init
